refactor(scripts): file-organization sweep over the constantScope and repeatedListItems scans
The two scans' functions, tests and the constant model move into the layers, and scanCode's yield gets a named tuple type. getModuleScopeConstants read a k-line declaration k times over — rescanning a growing prefix once per line to find its end, then scanning the whole declaration again to classify it — and now makes one lazy pass that stops at the first depth-zero semicolon and keeps the tokens it collected. The scan's output over the repository is byte-identical before and after.
